+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 **Goal: Offer actions when runtime errors occur.**
 **Goal: Separate line/column in go-to-def input positions.**
 
-# 0.20 (unreleased)
+# 0.20 (released 1st June 2025)
 **Goal: Namespaces.**
 
 ### Diagnostics